Move Qc2Formed handling to QuorumProposalRecvTaskState to unblock ConsensusTaskState processing
