+++ hotshot-task-impls/src/consensus/mod.rs
@@ -7,7 +7,6 @@
 use async_broadcast::{Receiver, Sender};
 use async_lock::RwLock;
 use async_trait::async_trait;
-use either::Either;
 use hotshot_task::task::TaskState;
 use hotshot_types::{
     consensus::OuterConsensus,
@@ -19,7 +18,6 @@ use hotshot_types::{
         node_implementation::{NodeImplementation, NodeType, Versions},
         signature_key::SignatureKey,
     },
-    utils::option_epoch_from_block_number,
     vote::HasViewNumber,
 };
 use hotshot_utils::anytrace::*;
@@ -30,7 +28,7 @@ use tracing::instrument;
 use self::handlers::{
     handle_quorum_vote_recv, handle_timeout, handle_timeout_vote_recv, handle_view_change,
 };
-use crate::helpers::{validate_qc_and_next_epoch_qc, wait_for_next_epoch_qc};
+use crate::helpers::validate_qc_and_next_epoch_qc;
 use crate::{events::HotShotEvent, helpers::broadcast_event, vote_collection::VoteCollectorsMap};
 
 /// Event handlers for use in the `handle` method.
@@ -139,59 +137,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>, V: Versions> ConsensusTaskSt
                 if let Err(e) = handle_timeout(*view_number, *epoch, &sender, self).await {
                     tracing::debug!("Failed to handle Timeout event; error = {e}");
                 }
-            }
-            HotShotEvent::Qc2Formed(Either::Left(quorum_cert)) => {
-                let cert_view = quorum_cert.view_number();
-                if !self.upgrade_lock.epochs_enabled(cert_view).await {
-                    tracing::debug!("QC2 formed but epochs not enabled. Do nothing");
-                    return Ok(());
-                }
-                if !self
-                    .consensus
-                    .read()
-                    .await
-                    .is_leaf_extended(quorum_cert.data.leaf_commit)
-                {
-                    tracing::debug!("We formed QC but not eQC. Do nothing");
-                    return Ok(());
-                }
-                if wait_for_next_epoch_qc(
-                    quorum_cert,
-                    &self.consensus,
-                    self.timeout,
-                    self.view_start_time,
-                    &receiver,
-                )
-                .await
-                .is_none()
-                {
-                    tracing::warn!("We formed eQC but we don't have corresponding next epoch eQC.");
-                    return Ok(());
-                }
-                let cert_block_number = self
-                    .consensus
-                    .read()
-                    .await
-                    .saved_leaves()
-                    .get(&quorum_cert.data.leaf_commit)
-                    .context(error!(
-                        "Could not find the leaf for the eQC. It shouldn't happen."
-                    ))?
-                    .height();
-
-                let cert_epoch = option_epoch_from_block_number::<TYPES>(
-                    true,
-                    cert_block_number,
-                    self.epoch_height,
-                );
-                // Transition to the new epoch by sending ViewChange
-                let next_epoch = cert_epoch.map(|x| x + 1);
-                tracing::info!("Entering new epoch: {:?}", next_epoch);
-                broadcast_event(
-                    Arc::new(HotShotEvent::ViewChange(cert_view + 1, next_epoch)),
-                    &sender,
-                )
-                .await;
             }
             HotShotEvent::ExtendedQcRecv(high_qc, next_epoch_high_qc, _) => {
                 if !self
