+++ hotshot-task-impls/src/helpers.rs
@@ -876,10 +876,12 @@ pub async fn wait_for_next_epoch_qc<TYPES: NodeType>(
     // TODO configure timeout
     let Some(time_spent) = Instant::now().checked_duration_since(view_start_time) else {
         // Shouldn't be possible, now must be after the start
+        tracing::error!("lrzasik: Shouldn't be possible, now must be after the start");
         return None;
     };
     let Some(time_left) = wait_duration.checked_sub(time_spent) else {
         // No time left
+        tracing::error!("lrzasik: No time left");
         return None;
     };
     let receiver = receiver.clone();
@@ -907,10 +909,12 @@ pub async fn wait_for_next_epoch_qc<TYPES: NodeType>(
                 return Some(next_epoch_qc.clone());
             }
         };
+        tracing::error!("lrzasik: timeout elapsed");
         return None;
     };
     let HotShotEvent::NextEpochQc2Formed(Either::Left(next_epoch_qc)) = event.as_ref() else {
         // this shouldn't happen
+        tracing::error!("lrzasik: event mismatch. this shouldn't happen");
         return None;
     };
     Some(next_epoch_qc.clone())
