+++ hotshot/src/tasks/task_state.rs
@@ -300,6 +300,7 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>, V: Versions> CreateTaskState
             id: handle.hotshot.id,
             upgrade_lock: handle.hotshot.upgrade_lock.clone(),
             epoch_height: handle.hotshot.config.epoch_height,
+            view_start_time: Instant::now(),
         }
     }
 }
